moved a lot of stuff around, modified a bit the RL agents and DC controller, they are probably broken for now.
