+++ omniisaacgymenvs/ros/ros_nodes.py
@@ -10,11 +10,11 @@ from std_msgs.msg import ByteMultiArray
 from geometry_msgs.msg import PoseStamped, Point, Pose
 
 from ros.ros_utills import derive_velocities
-from omniisaacgymenvs.mujoco_envs.position_controller_RL import PositionController
-from omniisaacgymenvs.mujoco_envs.pose_controller_RL import PoseController
-from omniisaacgymenvs.mujoco_envs.linear_velocity_tracker_RL import VelocityTracker, TrajectoryTracker
-from omniisaacgymenvs.mujoco_envs.pose_controller_DC import PoseController as PoseControllerDC
-from omniisaacgymenvs.mujoco_envs.RL_games_model_4_mujoco import RLGamesModel
+from omniisaacgymenvs.mujoco_envs.legacy.position_controller_RL import PositionController
+from omniisaacgymenvs.mujoco_envs.legacy.pose_controller_RL import PoseController
+from omniisaacgymenvs.mujoco_envs.legacy.linear_velocity_tracker_RL import VelocityTracker, TrajectoryTracker
+from omniisaacgymenvs.mujoco_envs.legacy.pose_controller_DC import PoseController as PoseControllerDC
+from omniisaacgymenvs.mujoco_envs.controllers.RL_games_model_4_mujoco import RLGamesModel
 
 class RLPlayerNode:
     def __init__(self, model: RLGamesModel, task_id: int, exp_settings, map:List[int]=[2,5,4,7,6,1,0,3]) -> None:
